+++ cms/urls.py
@@ -1,11 +1,9 @@
-import debug_toolbar
 from django.conf.urls import include, re_path
 from django.contrib import admin
 from django.urls import path
 from django.views.generic.base import TemplateView
 
 urlpatterns = [
-    re_path(r"^__debug__/", include(debug_toolbar.urls)),
     path(
         "robots.txt",
         TemplateView.as_view(template_name="robots.txt", content_type="text/plain"),